lock files with negative versions

+++ MakonFM/lib/MakonFM/Controller/SubSubmit.pm
@@ -1,6 +1,7 @@
 package MakonFM::Controller::SubSubmit;
 use Moose;
 use namespace::autoclean;
+use utf8;
 use Encode qw(encode_utf8);
 use MakonFM::Util::MatchChunk;
 use MakonFM::Util::Vyslov;
@@ -21,7 +22,7 @@ sub index :Path :Args(0) {
         $c->response->body('');
         $c->detach;
     }
-    
+
     my $param = $c->request->parameters;
     if ($c->request->user_agent =~ /MSIE/
         and keys(%{ $c->request->parameters }) == 0
@@ -31,53 +32,64 @@ sub index :Path :Args(0) {
         my $body = join '', <$handle>;
         $param = URL::Encode::url_params_mixed($body);
     }
-    
+
     my $trans = $param->{trans};
     my $trans_bytes = encode_utf8 $trans;
     my $filestem = $param->{filestem};
-    my $mfcc_fn = 
-        $c->econf(qw{paths audio mfcc})
-        . $filestem
-        . '.mfcc'
-    ;
+
     my $start = $param->{start};
     my $end   = $param->{ end };
-    
-    my $mp3_fn;
-    if ($c->config->{mkwav}) {
-        $mp3_fn =
-            $c->econf(qw{paths audio mp3})
+    my $subs;
+    my $is_success;
+
+    my $version = $c->model->resultset('Version')->find({key => $filestem})->get_column('value');
+    if ($version < 0) {
+        $is_success = 0;
+    }
+    else {
+        my $mfcc_fn =
+            $c->econf(qw{paths audio mfcc})
             . $filestem
-            . '.mp3'
+            . '.mfcc'
         ;
-    }
-    
-    MakonFM::Util::Vyslov::set_dict($c->model->resultset('Dict'));
 
-    undef $@;
-    my $subs = eval {
-        MakonFM::Util::MatchChunk::get_subs(\$trans_bytes, $mfcc_fn, $start, $end, $mp3_fn);
-    };
-    if ($@) {
-        $c->response->status(400);
-        $c->response->body(JSON->new->encode({message=>$@}));
-        $c->detach();
+        my $mp3_fn;
+        if ($c->config->{mkwav}) {
+            $mp3_fn =
+                $c->econf(qw{paths audio mp3})
+                . $filestem
+                . '.mp3'
+            ;
+        }
+
+        MakonFM::Util::Vyslov::set_dict($c->model->resultset('Dict'));
+
+        undef $@;
+        $subs = eval {
+            MakonFM::Util::MatchChunk::get_subs(\$trans_bytes, $mfcc_fn, $start, $end, $mp3_fn);
+        };
+        if ($@) {
+            $c->response->status(400);
+            $c->response->body(JSON->new->encode({message=>$@}));
+            $c->detach();
+        }
+
+        $is_success = $subs->{success};
+        $subs->{filestem} = $filestem;
+        $subs->{start} = $start;
+        $subs->{end} = $end;
+        $_->{humanic} = 1 for @{ $subs->{data} };
     }
 
-    $subs->{filestem} = $filestem;
-    $subs->{start} = $start;
-    $subs->{end} = $end;
-    $_->{humanic} = 1 for @{ $subs->{data} };
-    
     my $author = $param->{'author'} || '';
     my $session = $param->{session} || $c->sessionid || '';
-    
+
     $c->model->resultset('Submission')->create({
         filestem => $filestem,
         start_ts => $start,
         end_ts   => $end,
         transcription => $trans_bytes,
-        matched_ok => $subs->{success},
+        matched_ok => $is_success,
         sub_infos => [
             {
                 name => 'ip_address',
@@ -97,12 +109,23 @@ sub index :Path :Args(0) {
             },
         ],
     });
-    
-    if ($subs->{success}) {
+
+    if ($is_success) {
         MakonFM::Util::Subs::merge($subs);
     }
-    
-    $c->response->body(JSON->new->pretty->encode($subs));
+
+    if ($subs) {
+        $c->response->body(JSON->new->pretty->encode($subs));
+    }
+    else {
+        $c->response->status(403);
+        $c->response->body(JSON->new->encode({
+            message => 'Tuto nahrávku již nelze opravovat. Pokud jste našli chybu v přepisu, prosím o zprávu.',
+        }));
+    }
+}
+
+sub save_submission {
 }
 
 __PACKAGE__->meta->make_immutable;
